Expand the egos bitfield to allow for the 32nd ego (lunging). [It was this or pick one to remove to allow lunging to exist.] Surprisingly easy, but increases savefile size substantially.

+++ crawl-ref/source/player.h
@@ -323,7 +323,7 @@ public:
     int magic_contamination;
 
     FixedVector<uint32_t, NUM_WEAPONS> seen_weapon;
-    FixedVector<uint32_t, NUM_ARMOURS> seen_armour;
+    FixedVector<uint64_t, NUM_ARMOURS> seen_armour;
     FixedBitVector<NUM_MISCELLANY>     seen_misc;
     uint8_t                            octopus_king_rings;
